Disable xdebug and enable gzip compression
XDebug was causing unrealistic performance numbers and was never
actually configured anyway. GZIP compression helps with over-the-wire
file sizes.

This was done after trying to work out why the customer listing endpoint
was taking upwards of 300ms. It was a dodgy index (fixed and merged in
to the previous commit with an amend; specifically the updated_at index
was doing more harm than good until the company_id was also part of the
index, in ascending order) and xdebug that were the culprit.