Allow resource IDs to be generated client-side
In order to achieve this securely the ID must now only be unique to the
specific company so if someone messes with the client-side code to try
and wreak havoc with non-unique IDs then it only has an effect on the
company the user is logged in for.

In addition company and user IDs are still generated server-side because
these must be unique.

The end effect of this is that now the resources can be stored and used
client-side, even if an internet connection isn't available to save the
data on the server,