Measurement settings model/migration (unoptimized)
Because I want to use postgres to its full potential I've had to create
some extensions to the postgres connection/grammar that comes with
laravel. Although this might seem a little overkill at first it's going
to become very useful once the enum types are used in several locations
in the DB. It's much better than having pointless tables with foreign
keys, especially when a many-to-many relationship will be required and
all we want to do is have a list of values rather than a list of
entities.