Measurement profile tables and save body profile
Create the measurement profile table/model structure and create a
default body profile when creating a new customer.

The customer can only ever have one body profile so it makes sense to
create it immediately when the customer is created meaning that all API
requests for creating new profiles can only create garment profiles.
Further, down the line when adding support for deleting profiles, only
garment measurement profiles can be deleted (soft deleted as they may be
referenced on orders).

One thing to note is that aside from deleting a profile all data is
immutable with changes going in a separate 'commits' table that build up
the current state of a profile.

Another thing to note is that for the body profile the garment column
must be null. The name and sample garment on each commit is also ignored
(and kept null at the application layer).