Create measurement profile commits
This doesn't include creating measurement profiles so currently the only
profile that can be committed to is the BODY profile which is created
when a customer is created.

Measurement profiles along with their full commit history and the
current measurement values are returned when GETting a customer.

Further, fixed issues with composite primary keys which aren't actually
supported by Laravel without the use of a plugin.